Move legacy log symlink to kuberuntime
Also remove the dockertools.DockerType constant.
