+++ pkg/kubelet/dockertools/docker_test.go
@@ -18,14 +18,10 @@ package dockertools
 
 import (
 	"encoding/json"
-	"fmt"
-	"math/rand"
-	"path"
 	"strings"
 	"testing"
 
 	"github.com/docker/docker/pkg/jsonmessage"
-	"github.com/stretchr/testify/assert"
 	"k8s.io/kubernetes/pkg/api/v1"
 	"k8s.io/kubernetes/pkg/credentialprovider"
 	"k8s.io/kubernetes/pkg/kubelet/dockershim/libdocker"
@@ -180,25 +176,4 @@ func TestPullWithSecrets(t *testing.T) {
 			t.Errorf("images pulled do not match the expected: %v", err)
 		}
 	}
-}
-
-const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-
-func randStringBytes(n int) string {
-	b := make([]byte, n)
-	for i := range b {
-		b[i] = letterBytes[rand.Intn(len(letterBytes))]
-	}
-	return string(b)
-}
-
-func TestLogSymLink(t *testing.T) {
-	as := assert.New(t)
-	containerLogsDir := "/foo/bar"
-	podFullName := randStringBytes(128)
-	containerName := randStringBytes(70)
-	dockerId := randStringBytes(80)
-	// The file name cannot exceed 255 characters. Since .log suffix is required, the prefix cannot exceed 251 characters.
-	expectedPath := path.Join(containerLogsDir, fmt.Sprintf("%s_%s-%s", podFullName, containerName, dockerId)[:251]+".log")
-	as.Equal(expectedPath, LogSymlink(containerLogsDir, podFullName, containerName, dockerId))
 }
