+++ pkg/kubelet/kuberuntime/legacy.go
@@ -17,8 +17,10 @@ limitations under the License.
 package kuberuntime
 
 import (
+	"fmt"
+	"path"
+
 	kubecontainer "k8s.io/kubernetes/pkg/kubelet/container"
-	"k8s.io/kubernetes/pkg/kubelet/dockertools"
 )
 
 // This file implements the functions that are needed for backward
@@ -30,12 +32,24 @@ const (
 	// kubelet.containerLogsDir.
 	legacyContainerLogsDir = "/var/log/containers"
 	// legacyLogSuffix is the legacy log suffix.
-	legacyLogSuffix = dockertools.LogSuffix
+	legacyLogSuffix = "log"
+
+	ext4MaxFileNameLen = 255
 )
 
 // legacyLogSymlink composes the legacy container log path. It is only used for legacy cluster
 // logging support.
 func legacyLogSymlink(containerID string, containerName, podName, podNamespace string) string {
-	return dockertools.LogSymlink(legacyContainerLogsDir, kubecontainer.BuildPodFullName(podName, podNamespace),
+	return logSymlink(legacyContainerLogsDir, kubecontainer.BuildPodFullName(podName, podNamespace),
 		containerName, containerID)
+}
+
+func logSymlink(containerLogsDir, podFullName, containerName, dockerId string) string {
+	suffix := fmt.Sprintf(".%s", legacyLogSuffix)
+	logPath := fmt.Sprintf("%s_%s-%s", podFullName, containerName, dockerId)
+	// Length of a filename cannot exceed 255 characters in ext4 on Linux.
+	if len(logPath) > ext4MaxFileNameLen-len(suffix) {
+		logPath = logPath[:ext4MaxFileNameLen-len(suffix)]
+	}
+	return path.Join(containerLogsDir, logPath+suffix)
 }
