+++ pkg/kubelet/sysctl/runtime.go
@@ -21,7 +21,6 @@ import (
 
 	v1helper "k8s.io/kubernetes/pkg/api/v1/helper"
 	"k8s.io/kubernetes/pkg/kubelet/container"
-	"k8s.io/kubernetes/pkg/kubelet/dockertools"
 	"k8s.io/kubernetes/pkg/kubelet/lifecycle"
 )
 
@@ -31,7 +30,12 @@ const (
 	// (e.g., 1.24). Append the version with a ".0" so that it works
 	// with both the CRI and dockertools comparison logic.
 	dockerMinimumAPIVersion = "1.24.0"
+
+	dockerTypeName = "docker"
 )
+
+// TODO: The admission logic in this file is runtime-dependent. It should be
+// changed to be generic and CRI-compatible.
 
 type runtimeAdmitHandler struct {
 	result lifecycle.PodAdmitResult
@@ -42,7 +46,7 @@ var _ lifecycle.PodAdmitHandler = &runtimeAdmitHandler{}
 // NewRuntimeAdmitHandler returns a sysctlRuntimeAdmitHandler which checks whether
 // the given runtime support sysctls.
 func NewRuntimeAdmitHandler(runtime container.Runtime) (*runtimeAdmitHandler, error) {
-	if runtime.Type() == dockertools.DockerType {
+	if runtime.Type() == dockerTypeName {
 		v, err := runtime.APIVersion()
 		if err != nil {
 			return nil, fmt.Errorf("failed to get runtime version: %v", err)
